conversion,remote: Bring old v2 code back to prepare v3
Only cut & paste. It doesn't even compile yet.
